+++ onnxruntime/test/onnx/microbenchmark/quantize.cc
@@ -3,6 +3,7 @@
 #include <benchmark/benchmark.h>
 #include "core/util/qmath.h"
 #include "core/util/thread_utils.h"
+#include "core/framework/int4.h"
 
 static void BenchSize(benchmark::internal::Benchmark* b) {
   for (int size : {80000, 160000, 320000, 640000, 1280000}) {
@@ -77,3 +78,104 @@ BENCHMARK(BM_Quantize)
     ->UseRealTime()
     ->Unit(benchmark::TimeUnit::kNanosecond)
     ->Apply(BenchSize);
+
+static void BM_BlockedQuantize_NotLastAxis(benchmark::State& state) {
+  using Int4 = onnxruntime::Int4x2;
+  using UnpackedType = Int4::UnpackedType;
+  const std::ptrdiff_t M[] = {96, 192, 192};
+  const std::ptrdiff_t N[] = {2048, 2048, 4096};
+  const int64_t size_idx = state.range(0);
+  const int64_t threads = state.range(1);
+  const int64_t block_size = state.range(2);
+  size_t batch_size = M[size_idx] * N[size_idx];
+  size_t quant_block_size = 64;
+  size_t scale_size = batch_size / quant_block_size;
+
+  float* a_data = GenerateArrayWithRandomValue<float>(batch_size, -16, 14);
+  size_t a_quant_size = sizeof(Int4::UnpackedType) * Int4::CalcNumInt4Pairs(batch_size);
+  float* scale = GenerateArrayWithRandomValue<float>(scale_size, 1.95f, 2.33f);
+  UnpackedType* zero_point = GenerateArrayWithRandomValue<UnpackedType>(Int4::CalcNumInt4Pairs(scale_size), -1, 1);
+  UnpackedType* a_data_quant = static_cast<UnpackedType*>(aligned_alloc(a_quant_size, 64));
+
+  OrtThreadPoolParams tpo;
+  tpo.thread_pool_size = int(threads);
+  tpo.auto_set_affinity = true;
+  std::unique_ptr<onnxruntime::concurrency::ThreadPool> tp(
+      onnxruntime::concurrency::CreateThreadPool(&onnxruntime::Env::Default(),
+                                                 tpo, onnxruntime::concurrency::ThreadPoolType::INTRA_OP));
+
+  for (auto _ : state) {
+    benchmark::DoNotOptimize(a_data_quant);
+    onnxruntime::BlockedQuantizeLinear<float, Int4, 2>::opNotLastAxis(
+        tp.get(), a_data, scale, reinterpret_cast<Int4*>(zero_point), reinterpret_cast<Int4*>(a_data_quant),
+        1, M[size_idx], N[size_idx], static_cast<std::ptrdiff_t>(quant_block_size), static_cast<std::ptrdiff_t>(block_size), true);
+    benchmark::ClobberMemory();
+  }
+  aligned_free(a_data_quant);
+  aligned_free(a_data);
+  aligned_free(scale);
+  aligned_free(zero_point);
+}
+
+static void BM_BlockedQuantize_LastAxis(benchmark::State& state) {
+  using Int4 = onnxruntime::Int4x2;
+  using UnpackedType = Int4::UnpackedType;
+  const std::ptrdiff_t M[] = {96, 192, 192};
+  const std::ptrdiff_t N[] = {2048, 2048, 4096};
+  const int64_t size_idx = state.range(0);
+  const int64_t threads = state.range(1);
+  const int64_t quant_block_size = state.range(2);
+  size_t batch_size = M[size_idx] * N[size_idx];
+  size_t scale_size = batch_size / quant_block_size;
+
+  float* a_data = GenerateArrayWithRandomValue<float>(batch_size, -16, 14);
+  size_t a_quant_size = sizeof(Int4::UnpackedType) * Int4::CalcNumInt4Pairs(batch_size);
+  float* scale = GenerateArrayWithRandomValue<float>(scale_size, 1.95f, 2.33f);
+  UnpackedType* zero_point = GenerateArrayWithRandomValue<UnpackedType>(Int4::CalcNumInt4Pairs(scale_size), -1, 1);
+  UnpackedType* a_data_quant = static_cast<UnpackedType*>(aligned_alloc(a_quant_size, 64));
+
+  OrtThreadPoolParams tpo;
+  tpo.thread_pool_size = int(threads);
+  tpo.auto_set_affinity = true;
+  std::unique_ptr<onnxruntime::concurrency::ThreadPool> tp(
+      onnxruntime::concurrency::CreateThreadPool(&onnxruntime::Env::Default(),
+                                                 tpo, onnxruntime::concurrency::ThreadPoolType::INTRA_OP));
+
+  for (auto _ : state) {
+    benchmark::DoNotOptimize(a_data_quant);
+    onnxruntime::BlockedQuantizeLinear<float, Int4, 2>::opLastAxis(
+        tp.get(), a_data, scale, reinterpret_cast<Int4*>(zero_point), reinterpret_cast<Int4*>(a_data_quant),
+        M[size_idx], N[size_idx], static_cast<std::ptrdiff_t>(quant_block_size), true);
+    benchmark::ClobberMemory();
+  }
+  aligned_free(a_data_quant);
+  aligned_free(a_data);
+  aligned_free(scale);
+  aligned_free(zero_point);
+}
+
+BENCHMARK(BM_BlockedQuantize_NotLastAxis)
+    ->UseRealTime()
+    ->Unit(benchmark::TimeUnit::kNanosecond)
+    ->Apply([](benchmark::internal::Benchmark* b) {
+      for (int size_idx : {0, 1, 2}) {
+        for (int thread : {2, 4, 8}) {
+          for (int block_size : {64, 128}) {
+            b->Args({size_idx, thread, block_size});
+          }
+        }
+      }
+    });
+
+BENCHMARK(BM_BlockedQuantize_LastAxis)
+    ->UseRealTime()
+    ->Unit(benchmark::TimeUnit::kNanosecond)
+    ->Apply([](benchmark::internal::Benchmark* b) {
+      for (int size_idx : {0, 1, 2}) {
+        for (int thread : {2, 4, 8}) {
+          for (int quant_block_size : {16, 64, 256}) {
+            b->Args({size_idx, thread, quant_block_size});
+          }
+        }
+      }
+    });
